Refactored examples to include points of which the timestamp matches the start (rather than the end) of the index interval
Theoretically, this should result in the same tag. It is much more intuitive to think about, however, in the case where the custom calculation calculates UTC monthly block aggregates. Since the historic indexing also requests data in blocks of UTC months, the value for a month will now actually be calculated during the index request for that month (rather than for the month before).